+++ backend/app/models/court.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, Float, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -10,8 +10,8 @@ class RentalTemplate(Base):
     # the app runs against the existing production DB.
     __tablename__ = "rental_tamplate"
 
-    id = Column(Integer, primary_key=True)
-    club_id = Column(Integer, ForeignKey("club.id"))
+    id = Column(BigInteger, primary_key=True)
+    club_id = Column(BigInteger, ForeignKey("club.id"))
     start_effective_date = Column(Date)
     end_effective_date = Column(Date)
     days_str = Column(String(255))   # comma-separated day numbers 1-7 (Sun-Sat)
@@ -33,14 +33,14 @@ class RentalTemplate(Base):
 class AvailableCourtSlot(Base):
     __tablename__ = "available_courts_search"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     # Python attribute stays rental_template_id; DB column/FK use the prod typo.
-    rental_template_id = Column("rental_tamplate_id", Integer, ForeignKey("rental_tamplate.id"))
+    rental_template_id = Column("rental_tamplate_id", BigInteger, ForeignKey("rental_tamplate.id"))
     hour = Column(Integer)
     curdate = Column(Date)
     taken = Column(DateTime, nullable=True)
     is_holiday = Column(String(1), nullable=True)  # Y or null
-    order_id = Column(Integer, ForeignKey("court_order.id"), nullable=True)
+    order_id = Column(BigInteger, ForeignKey("court_order.id"), nullable=True)
 
     rental_template = relationship("RentalTemplate", back_populates="slots")
     order = relationship("CourtOrder", back_populates="slot")
@@ -49,8 +49,8 @@ class AvailableCourtSlot(Base):
 class CourtLockDate(Base):
     __tablename__ = "court_lock_dates"
 
-    id = Column(Integer, primary_key=True)
-    club_id = Column(Integer, ForeignKey("club.id"))
+    id = Column(BigInteger, primary_key=True)
+    club_id = Column(BigInteger, ForeignKey("club.id"))
     day_of_week = Column(Integer)
     from_hour = Column(Integer)
     to_hour = Column(Integer)
@@ -61,8 +61,8 @@ class CourtLockDate(Base):
 class HolidayDate(Base):
     __tablename__ = "holiday_dates"
 
-    id = Column(Integer, primary_key=True)
-    club_id = Column(Integer, ForeignKey("club.id"))
+    id = Column(BigInteger, primary_key=True)
+    club_id = Column(BigInteger, ForeignKey("club.id"))
     court_number = Column(Integer, nullable=True)  # NULL = all courts in the club
     start_date = Column(Date)
     end_date = Column(Date)
@@ -75,8 +75,8 @@ class HolidayDate(Base):
 class HolidayOverwrite(Base):
     __tablename__ = "holiday_overwrite"
 
-    id = Column(Integer, primary_key=True)
-    club_id = Column(Integer, ForeignKey("club.id"))
+    id = Column(BigInteger, primary_key=True)
+    club_id = Column(BigInteger, ForeignKey("club.id"))
     date = Column(Date)
     start_hour = Column(Integer)
     end_hour = Column(Integer)
