+++ backend/app/models/misc.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Column, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -8,7 +8,7 @@ from app.models.types import Date  # DATE column tolerant of production's DATETI
 class ResetPassword(Base):
     __tablename__ = "reset_password"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     username = Column(String(255))
     token = Column(String(255))
     date_created = Column(Date)
@@ -17,7 +17,7 @@ class ResetPassword(Base):
 class Contact(Base):
     __tablename__ = "contact"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     first_name = Column(String(255))
     last_name = Column(String(255))
     phone_number = Column(String(255), nullable=True)
@@ -27,7 +27,7 @@ class Contact(Base):
     type_of = Column(String(255))
     # Global "צור קשר" form has no club (messages go to the BACO inbox), so this
     # is nullable — the legacy schema had it NOT NULL (see migration 003).
-    club_id = Column(Integer, ForeignKey("club.id"), nullable=True)
+    club_id = Column(BigInteger, ForeignKey("club.id"), nullable=True)
 
     club = relationship("Club")
 
@@ -35,6 +35,6 @@ class Contact(Base):
 class PelecardErrorList(Base):
     __tablename__ = "pelecard_error_list"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     description = Column(String(500))
     error_code = Column(String(10))
