+++ backend/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Boolean, Column, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -7,14 +7,14 @@ from app.database import Base
 class Role(Base):
     __tablename__ = "role"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     authority = Column(String(255), unique=True, nullable=False)
 
 
 class User(Base):
     __tablename__ = "user"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     username = Column(String(255), unique=True, nullable=False)  # email address
     password = Column(String(255), nullable=False)
     first_name = Column(String(255), nullable=False)
@@ -33,8 +33,8 @@ class User(Base):
 class UserRole(Base):
     __tablename__ = "user_role"
 
-    user_id = Column(Integer, ForeignKey("user.id"), primary_key=True)
-    role_id = Column(Integer, ForeignKey("role.id"), primary_key=True)
+    user_id = Column(BigInteger, ForeignKey("user.id"), primary_key=True)
+    role_id = Column(BigInteger, ForeignKey("role.id"), primary_key=True)
 
     user = relationship("User", back_populates="roles")
     role = relationship("Role")
